Extract dynamic_fst code from replace
